Add EVM module
