+++ runtime/shibuya/src/lib.rs
@@ -69,8 +69,9 @@ use sp_runtime::{
 use sp_std::prelude::*;
 
 pub use astar_primitives::{
-    ethereum_checked::CheckedEthereumTransact, xcm::AssetLocationIdConverter, AccountId, Address,
-    AssetId, Balance, BlockNumber, EvmRevertCodeHandler, Hash, Header, Index, Signature,
+    ethereum_checked::CheckedEthereumTransact, evm::EvmRevertCodeHandler,
+    xcm::AssetLocationIdConverter, AccountId, Address, AssetId, Balance, BlockNumber, Hash, Header,
+    Index, Signature,
 };
 
 use pallet_evm_precompile_assets_erc20::AddressToAssetId;
